Refactor duplicated templating process into single function.
Create the setvars function to handle reading and settings variables.
While this simplifies the code in rss.sh and index.sh, it forces
assemble.sh to have to rely on a temporary file. If I can find a way to
refactor this in the future to avoid having to use the filesystem, then
I'll do so.