Refactor templating engine and generation of auxiliary files.
Expand on templating by providing the ARTICLES variable and an xtemplate
helper function that can be used in variable declarations. This enables
the RSS feed and index file to be generated by templates.

However, considerable complexity now exists and the templating code is
sorta confusing in places. Probably needs future simplification.